services/nft-software-engine/requirements.txt
# NFT Software Engine Requirements

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0  # For async SQLite support

# Web3 & Blockchain
web3==6.11.1
eth-account==0.9.0
hexbytes==0.3.1

# HTTP Client & API
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching & Rate Limiting
redis==5.0.1
aioredis==2.0.1
slowapi==0.1.9

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Environment & Configuration
python-dotenv==1.0.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx[test]==0.25.2

# Code Quality
black==23.9.1
isort==5.12.0
mypy==1.6.1
pylint==3.0.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.7
